添加玩家回合制的操作。添加通过划分操作码的不同范围，自定义多种消息。完善小人生成逻辑。添加镜头跟随

+++ MyServer/Server/Src/SimpleServer/Src/Msg/Msg.cs
@@ -68,7 +68,7 @@ public class PlayerServerInfo : BaseFormater {
 public class PlayerInput : BaseFormater
 {
 
-    public int number;
+    public short number;
     public float forceX;
     public float forceY;
 
@@ -84,7 +84,7 @@ public class PlayerInput : BaseFormater
 
     public override void Deserialize(Deserializer reader)
     {
-        number = reader.ReadInt32();
+        number = reader.ReadInt16();
         forceX = reader.ReadSingle();
         forceY = reader.ReadSingle();
